Fix invalid reporting during block return
A block containing a `return` keyword can cause return events to
report invalid return values. Consider the example below

    class Foo
      # @return [Integer]
      def bar
        baz { return 1 }
      end

      # @return [Symbol]
      def norf
        baz
      end

      private

      # @return [Symbol]
      def baz(&block)
        qux(&block) if block_given?

        :baz
      end

      # @return [Symbol]
      def qux
        yield if block_given?

        :qux
      end
    end

    Foo.new.bar  # => 1
    Foo.new.norf # => :baz

When you call `bar` you pass a block down through `baz` and `qux`. The
block is executed inside of `qux` via the `yield`. Since the original
block inside of the `bar` method contains a `return`, this `yield`
results in `qux` and `baz` terminating. According to `TracePoint`, these
methods are "returning" with the value `nil` but this return value is
impossible to actually use. Therefore, we track when a `b_return` event
occurs and flag the tracing state as ambiguous until this is resolved.